requirements.txt
asgiref==3.4.1
click==8.0.3
Django==3.2.9
django-crispy-forms==1.13.0
django-simple-chatbot==0.0.9
djangorestframework==3.12.4
joblib==1.1.0
nltk==3.6.5
Pillow==8.4.0
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
tqdm==4.62.3
